core (dsp): Use a single MDCT implementation using the fft module.
Since the fft module now supports SIMD via. RustFFT, MDCT no longer needs
a SIMD and non-SIMD implementation.
