+++ symphonia-core/src/dsp/mdct.rs
@@ -1,11 +1,13 @@
 // Symphonia
-// Copyright (c) 2019-2023 The Project Symphonia Developers.
+// Copyright (c) 2019-2022 The Project Symphonia Developers.
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//! The Modified Discrete Cosine Transform (MDCT) implemented without SIMD optimizations.
+//! The `mdct` module implements the Modified Discrete Cosine Transform (MDCT).
+//!
+//! The MDCT in this module is implemented in-terms of a forward FFT.
 
 use crate::dsp::complex::Complex;
 use crate::dsp::fft::*;
@@ -13,8 +15,7 @@ use crate::dsp::fft::*;
 /// The Inverse Modified Discrete Transform (IMDCT).
 pub struct Imdct {
     fft: Fft,
-    fft_in: Box<[Complex<f32>]>,
-    fft_out: Box<[Complex<f32>]>,
+    scratch: Box<[Complex<f32>]>,
     twiddle: Box<[Complex<f32>]>,
 }
 
@@ -36,7 +37,7 @@ impl Imdct {
         assert!(n.is_power_of_two(), "n must be a power of two");
         // A complex FFT of size N/2 is used to compute the IMDCT. Therefore, the maximum value of N
         // is 2 * Fft::MAX_SIZE.
-        assert!(n <= 2 * Fft::MAX_SIZE, "maximum size exceeded");
+        assert!(n <= 2 * MAX_SIZE, "maximum size exceeded");
 
         let n2 = n / 2;
         let mut twiddle = Vec::with_capacity(n2);
@@ -52,10 +53,10 @@ impl Imdct {
             twiddle.push(Complex::new(re as f32, im as f32));
         }
 
-        let fft_in = vec![Default::default(); n2].into_boxed_slice();
-        let fft_out = vec![Default::default(); n2].into_boxed_slice();
+        // Allocate scratch for the IMDCT.
+        let scratch = vec![Default::default(); n2].into_boxed_slice();
 
-        Imdct { fft: Fft::new(n2), fft_in, fft_out, twiddle: twiddle.into_boxed_slice() }
+        Imdct { fft: Fft::new(n2), scratch, twiddle: twiddle.into_boxed_slice() }
     }
 
     /// Performs the the N-point Inverse Modified Discrete Cosine Transform.
@@ -77,7 +78,7 @@ impl Imdct {
         assert_eq!(out.len(), 2 * n);
 
         // Pre-FFT twiddling and packing of the real input signal values into complex signal values.
-        for (i, (&w, t)) in self.twiddle.iter().zip(self.fft_in.iter_mut()).enumerate() {
+        for (i, (&w, t)) in self.twiddle.iter().zip(self.scratch.iter_mut()).enumerate() {
             let even = spec[i * 2];
             let odd = -spec[n - 1 - i * 2];
 
@@ -87,7 +88,7 @@ impl Imdct {
         }
 
         // Do the FFT.
-        self.fft.fft(&self.fft_in, &mut self.fft_out);
+        self.fft.fft_inplace(&mut self.scratch);
 
         // Split the output vector (2N samples) into 4 vectors (N/2 samples each).
         let (vec0, vec1) = out.split_at_mut(n2);
@@ -96,7 +97,7 @@ impl Imdct {
 
         // Post-FFT twiddling and processing to expand the N/2 complex output values into 2N real
         // output samples.
-        for (i, (x, &w)) in self.fft_out[..n4].iter().zip(self.twiddle[..n4].iter()).enumerate() {
+        for (i, (x, &w)) in self.scratch[..n4].iter().zip(self.twiddle[..n4].iter()).enumerate() {
             // The real and imaginary components of the post-twiddled FFT samples are used to
             // generate 4 reak output samples. Using the first half of the complex FFT output,
             // populate each of the 4 output vectors.
@@ -116,7 +117,7 @@ impl Imdct {
             vec3[fi] = val.re;
         }
 
-        for (i, (x, &w)) in self.fft_out[n4..].iter().zip(self.twiddle[n4..].iter()).enumerate() {
+        for (i, (x, &w)) in self.scratch[n4..].iter().zip(self.twiddle[n4..].iter()).enumerate() {
             // Using the second half of the FFT output samples, finish populating each of the 4
             // output vectors.
             let val = w * x.conj();
@@ -142,5 +143,60 @@ impl Imdct {
         // Additionally, note that vectors 0 and 3 are reversed copies (+ negation for vector 0) of
         // vectors 1 and 2, respectively. Pulling these copies out into a separate loop using
         // iterators yielded no measureable difference either.
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::f64;
+
+    fn imdct_analytical(x: &[f32], y: &mut [f32], scale: f64) {
+        assert!(y.len() == 2 * x.len());
+
+        // Generates 2N outputs from N inputs.
+        let n_in = x.len();
+        let n_out = x.len() << 1;
+
+        let pi_2n = f64::consts::PI / (2 * n_out) as f64;
+
+        for (i, item) in y.iter_mut().enumerate().take(n_out) {
+            let accum: f64 = x
+                .iter()
+                .copied()
+                .map(f64::from)
+                .enumerate()
+                .take(n_in)
+                .map(|(j, jtem)| jtem * (pi_2n * ((2 * i + 1 + n_in) * (2 * j + 1)) as f64).cos())
+                .sum();
+
+            *item = (scale * accum) as f32;
+        }
+    }
+
+    #[test]
+    fn verify_imdct() {
+        #[rustfmt::skip]
+        const TEST_VECTOR: [f32; 32] = [
+             1.0,  2.0,  3.0,  4.0,  5.0,  6.0,  7.0,  8.0,
+             9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0,
+            17.0, 18.0, 19.0, 20.0, 21.0, 22.0, 23.0, 24.0,
+            25.0, 26.0, 27.0, 28.0, 29.0, 30.0, 31.0, 32.0,
+        ];
+
+        let mut actual = [0f32; 64];
+        let mut expected = [0f32; 64];
+
+        let scale = (2.0f64 / 64.0).sqrt();
+
+        imdct_analytical(&TEST_VECTOR, &mut expected, scale);
+
+        let mut mdct = Imdct::new_scaled(32, scale);
+        mdct.imdct(&TEST_VECTOR, &mut actual);
+
+        for i in 0..64 {
+            let delta = f64::from(actual[i]) - f64::from(expected[i]);
+            assert!(delta.abs() < 0.00001);
+        }
     }
 }
